Being moving payer consolidation logic into directory, breaking up logic
Created a higher order connection wrapper func that allows each step in the data consolidation to be able to be executed and tested independently. Data consolidation will be happening upon import but I'm trying to set things up such that it's easy to execute, edit, and test different parts of the pipeline.