Task/fix addit crit dev collection/plrs 516 (#432)
* Split and rename QoA and AddCrit pipelines

* Update formateAdditionalCriteriaDoc func

* Split QoA + AddCrit doc generation for materialized views

* Extend add crit pipeline to return actual docs + delete formatting step in script

* Refactor combined data formatter to include array of additional criteria fields

* Change combined data to write to MATT_FIX

* Change combine lives util to source from MATT_FIX collection

* Source regional targeting data from MATT_FIX collection

* Change field name for additCrit data in combined data formatter

* Add match stage to top of each pipeline to screen out null addl criteria, policy link, access data

* Remove no longer needed isValidQualityAccessDoc check

* Comment back in materialization funcs in materializeNonLivesCollections

* Realize we def shouldn't persist dupe fields in additional criteria slice of organizations.treatmentPlans.history, at least not in core

* Correct materialize additional criteria pipeline to not depend on additional criteria dupe string fields

Can't use $replaceRoot on the additionalCriteriaData slice because we can't source the indication, regimen, etc. fields from the dupe fields in there for the materialized view because:
A) What would be the purpose of doing all the lookups for org, ind, etc. here if we discard them at the end of the pipeline with $replaceRoot? [elaborated on in point B]
B) we can't depend on those fields because they might fall out of date as time passes when indication, regimen -- any of those fields are updated; we should rather depend on the string values gotten from lookups on the organizationId and treatmentPlanId at the point of materialization
C) we're going to be removing those inner fields anyway from the additionalCriteriaData slice as per previous commit so they won't be able to be used

* Add comment to pipeline about projecting _id out

* Remove _MATT safety postfixes everywhere in prep for running materialization against real collections on staging

Co-authored-by: Jonathan Lin <jon.cp.lin@gmail.com>